src/dashboard/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
azure-identity==1.15.0
azure-servicebus==7.11.4
